refactor: move userdata to common/user file
